Moving merge code from test to main folder
